# apps/sync-api/openapi.yaml
openapi: 3.1.0
info:
  title: FieldEdge Sync API
  version: 0.1.0
  description: "Orchestrator between FieldEdge mobile devices and the central Qdrant
cluster. Authentication uses HS256 JWTs issued by `/auth/login` and
refreshed via `/auth/refresh`. All `/sync/*` endpoints require an
`Authorization: Bearer <access_token>` header carrying a token whose `type` claim is `access` (15-minute TTL). See docs/08-AUTH.md for the planned OIDC swap.
"
  contact:
    name: FieldEdge Hackathon Team
    url: https://github.com/choksi2212/field-edge
  license:
    name: Proprietary
servers:
- url: http://localhost:8000
  description: Local development
- url: https://field-edge-sync-api.onrender.com
  description: Production (Render)
...
paths:
...
  /sync/upload:
    post:
      tags:
      - sync
      summary: Upload a batch of points from a device
      description: 'Accepts a batch of points from a device and writes them to the

        central Qdrant collection. Each point is upserted by id; if the

        server already has a point with the same id and the same

        `vector_checksum` the result is idempotent (`accepted`). A

        differing checksum triggers conflict resolution by timestamp:

        if the local `local_updated_at` is older than the remote copy

        the point is `rejected_too_old`; otherwise the local payload

        wins (`conflict_resolved`, `local_wins`).

        '
      operationId: uploadPoints
      security:
      - DeviceBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UploadRequest'
      responses:
        '200':
          description: Batch processed. Check per-point status.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '413':
          description: Batch too large (max 100 points)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          $ref: '#/components/responses/ValidationError'
  /sync/pull:
    get:
      tags:
      - sync
      summary: Pull updates from the central cluster
      description: "Returns points updated after the given cursor, excluding writes\nmade by the calling device. The cursor\
        \ is opaque (base64 JSON\n`{server_time, last_id}`) \xE2\u20AC\u201D clients must echo back the\n`next_cursor` they\
        \ receive.\n"
      operationId: pullUpdates
      security:
      - DeviceBearer: []
      parameters:
      - name: since
        in: query
        required: false
        schema:
          type: string
          description: Opaque cursor returned by a previous pull/upload.
        example: eyJzZXJ2ZXJfdGltZSI6ICIyMDI2LTA5LTI0VDAxOjAwOjAwKzAwOjAwIiwgImxhc3RfaWQiOiAiYWJjMTIzIn0=
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          default: 100
          maximum: 500
          minimum: 1
      responses:
        '200':
          description: Cursor-scoped slice of remote points.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PullResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '422':
          $ref: '#/components/responses/ValidationError'
...
  /sync/wal/replay:
    post:
      tags:
      - sync
      summary: Replay a previously-uploaded batch after a crash (FR-080)
      description: 'Used when the device believes a `POST /sync/upload` succeeded

        but never received a response (network dropped mid-flight). The

        device re-sends the same `batch_id` + points; the server

        upserts idempotently by point id and returns the per-point

        result.

        '
      operationId: walReplay
      security:
      - DeviceBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WalReplayRequest'
      responses:
        '200':
          description: Replay processed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '413':
          description: Batch too large (max 100 points)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          $ref: '#/components/responses/ValidationError'
...
components:
  securitySchemes:
    DeviceBearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'HS256-signed JWT issued by /auth/login. The token

        carries the device_id as its `sub` claim and the literal

        string `access` as its `type` claim. Tokens expire after

        JWT_ACCESS_TTL_SECONDS (default 900s = 15 min); refresh via

        /auth/refresh. The v0 login endpoint accepts any 8+ char

        device_token -- OIDC swap replaces that with a real

        auth-code exchange.'
  responses:
    Unauthorized:
      description: Missing or invalid Bearer token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ValidationError:
      description: Request payload failed schema validation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/HTTPValidationError'
  schemas:
    PointPayload:
      type: object
      description: 'Per-point metadata stored alongside the vector. Bumping

        `schema_version` is how we evolve this payload without breaking

        old clients (the server writes whatever the device sends and

        stamps `server_version` on conflict resolution).

        '
      additionalProperties: false
      required:
      - photo_id
      - device_id
      - captured_at
      - project_id
      - file_path
      - local_updated_at
      - vector_checksum
      properties:
        schema_version:
          type: integer
          default: 1
          description: Payload schema version. v1 = initial.
        photo_id:
          type: string
          description: ULID of the photo
        device_id:
          type: string
        captured_at:
          type: string
          format: date-time
        lat:
          type:
          - number
          - 'null'
          format: double
        lng:
          type:
          - number
          - 'null'
          format: double
        gps_status:
          type: string
          enum:
          - ok
          - unavailable
          - denied
          default: ok
        project_id:
          type: string
        file_path:
          type: string
          description: "Relative path on the device (e.g. `photos/01HF\xE2\u20AC\xA6jpg`)"
        embedding_status:
          type: string
          enum:
          - ok
          - pending
          - failed
          default: ok
        enrichment_id:
          type:
          - string
          - 'null'
        enrichment_tags:
          type: array
          items:
            type: string
          default: []
        enrichment_objects:
          type: array
          items:
            type: object
            additionalProperties: true
          default: []
        enrichment_text:
          type:
          - string
          - 'null'
        synced_at:
          type:
          - string
          - 'null'
          format: date-time
        local_updated_at:
          type: string
          format: date-time
          description: LWW timestamp on the device; used for conflict resolution.
        vector_checksum:
          type: string
          description: 'Hash of the embedding vector. Same value on retry = idempotent.

            '
    Point:
      type: object
      additionalProperties: false
      required:
      - id
      - vector
      - payload
      properties:
        id:
          type: string
          description: Point id (UUID for Qdrant Cloud compatibility)
        vector:
          type: array
          items:
            type: number
            format: float
          minItems: 512
          maxItems: 512
          description: 512-dim embedding (MobileNetV3 / CLIP variant)
        payload:
          $ref: '#/components/schemas/PointPayload'
    UploadRequest:
      type: object
      additionalProperties: false
      required:
      - device_id
      - batch_id
      - points
      properties:
        device_id:
          type: string
        batch_id:
          type: string
          description: Client-generated ULID for the upload batch
        points:
          type: array
          items:
            $ref: '#/components/schemas/Point'
          maxItems: 100
...
    WalReplayRequest:
      type: object
      additionalProperties: false
      required:
      - device_id
      - batch_id
      - points
      properties:
        device_id:
          type: string
        batch_id:
          type: string
        points:
          type: array
          items:
            $ref: '#/components/schemas/Point'
          maxItems: 100
        replay:
          type: boolean
          default: true
    PointResult:
      type: object
      additionalProperties: false
      required:
      - id
      - status
      properties:
        id:
          type: string
        status:
          type: string
          enum:
          - accepted
          - accepted_with_merge
          - conflict_resolved
          - rejected_too_old
          - rejected_invalid_payload
        resolution:
          type:
          - string
          - 'null'
          enum:
          - local_wins
          - remote_wins
          - merged
          - null
        resolved_payload:
          type:
          - object
          - 'null'
          additionalProperties: true
        error_message:
          type:
          - string
          - 'null'
        server_version:
          type:
          - integer
          - 'null'
          description: 'Monotonic per-point version on the server. Bumped on each

            conflict resolution.

            '
    UploadResponse:
      type: object
      additionalProperties: false
      required:
      - batch_id
      - server_time
      - results
      - next_cursor
      properties:
        batch_id:
          type: string
        server_time:
          type: string
          format: date-time
        results:
          type: array
          items:
            $ref: '#/components/schemas/PointResult'
        next_cursor:
          type: string
          description: Opaque cursor for the next pull.
    PullResponse:
      type: object
      additionalProperties: false
      required:
      - server_time
      - points
      - next_cursor
      - has_more
      properties:
        server_time:
          type: string
          format: date-time
        points:
          type: array
          items:
            type: object
            additionalProperties: true
            description: "Loose shape \xE2\u20AC\u201D the server returns `dict[str, Any]`. Real\nrows include `id`, `vector`\
              \ (often empty for scrolled\nresults), and `payload`.\n"
        next_cursor:
          type: string
        has_more:
          type: boolean
...
    ErrorResponse:
      type: object
      additionalProperties: true
      description: Generic FastAPI HTTP error envelope.
      properties:
        detail:
          type: string
    HTTPValidationError:
      type: object
      additionalProperties: false
      required:
      - detail
      properties:
        detail:
          type: array
          items:
            $ref: '#/components/schemas/ValidationErrorItem'
    ValidationErrorItem:
      type: object
      additionalProperties: true
      properties:
        loc:
          type: array
          items:
            oneOf:
            - type: string
            - type: integer
        msg:
          type: string
        type:
          type: string
...